Create UI and API folders for tests. Moved respective tests there. Updated test case names according to convention.
