+++ e2e/cypress/e2e/ui/cookie-modal.cy.js
